fix(#25): use form el to reset file input after added-files
BREAKING CHANGE: The hidden file input is now wrapped in a form element (potentially breaking some more custom layouts if the file input was used in a way apart from the default). The `fileInputStyles` data attr is renamed to `formStyles` and placed on the form wrapper element instead of the file input.
- Use a form wrapper element to reliably reset the file input after adding files. This will allow browsers to upload the same file twice, where previously the `change` event would not fire due to the filename not changing.